Add a test framework and initial set of tests
Add a test framework that creates a podman container that looks a bit
like a toolbox container, and runs tests inside there, with a mocked
flatpak-spawn on the path. Tests can extend the mocks for flatpak-spawn.

Add an initial set of tests that should test most of code.sh.

Add a CONTRIBUTING.md with information about coding style and documentation
about how to run tests and write tests.